Update playcontroller_update.sh
Unfortunately the previous version of the script blocks the PC programming in the future if the GPIOs are not reseted at the end of the procedure. 
So, if we have installed an update without reseting we now have to reset the GPIOs first, before proceeding with the usual PC update.